Initial commit of the Notion Pocket project
This commit includes the following files:
- server.js
- routes/pocket.js
- routes/notion.js
- validation/validation.js
- utils/utils.js
- test/pocket.test.js
- test/notion.test.js
- test/validation.test.js
- package.json
- .gitignore

This project is a system to query the Pocket APIs, pull the latest saves, and then add them to a notion database item using the Notion.so APIs.
It includes a Node.js server with Express.js and routes for the Pocket and Notion APIs, as well as validation and utility functions.
It also includes tests for the routes and functions using Jest.